Add files via upload
This last file, is a backup for the whole db. In order to create it, I got to Administration >> Data Export >> And on this step, I selected to create a backup only for "tables", "Dump data only", and also " Export to Dump project folder". In this one file I joined the backup for all of the tables.